orchestration/requirements.txt
langgraph>=0.1.0
langchain-core>=0.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
pinecone>=3.0.0
asyncpg>=0.29.0
fastapi>=0.110.0
uvicorn>=0.29.0